tidy vm code
